refactor: extract GeoservicesImportService partials by concern (#1144)
Main file 2,497 → 311 LOC (-88%). Partials by concern:
- Features.cs (FeatureService-specific scanning)
- Fidelity.cs (capability/metadata fidelity checks)
- Geometry.cs (geometry parsing & SRID handling)
- Helpers.cs (shared utility helpers)
- ImportSteps.cs (import step orchestration)
- Inventory.cs (service/layer/coverage inventory building)
- Metadata.cs (layer/field metadata extraction)
